refactor(ale_claw): Tier 4 — split memory-flush policy out of session.py
Moves the memory-flush POLICY (7 constants + should_run_memory_flush /
has_already_flushed_for_current_compaction) to memory_flush_policy.py, leaving
the flush EXECUTION in memory_flush.py. The two policy fns reference SessionState
only as a type annotation, so the new module uses `from __future__ import
annotations` + a TYPE_CHECKING import — no runtime dependency on session, so
session.py's re-export of all 9 symbols is cycle-free. Every existing
`from .session import ...` (__init__.py, agent_loop.py, tests) keeps working.
Pure move, no behavior change; full suite green (928 passed, 3 skipped).
